!== NEW UPDATE AVAILABLE: v1.0.1-ALPHA !==
* Added new option: enable-serverlist-motd
* added new option: motd-serverlist-message
* You can now customize your server list MOTD. (Untested)
* Added new option: enable-player-motd.
* Renamed Server-name to new-player-motd and old-player-motd.
It's explainable in the configurations file. We added a system, where it'll detect if the player is old or new. If the player's old (as in joined before), then it'll display a different MOTD compared to the new-player-motd (if user specifies). Old means the player's joined before, therefore, it'll show the joined before MOTD message, and new means the player's never joined before, which again, it'll show a different MOTD message compared to the old-player-motd message.
* Added new task file for MOTD messages.
* default-server-name may/may not work, so best to keep this option turned on for now. You may turn it off if you wish, but there may be bugs. Use it at your own risk!
* Fixed some syntax errors during coding phrases.
* Removed useless use import: pocketmine\Server; as it wasn't needed nor intented to be used due to being able to use $this->plugin (as in the main class) rather than Server::getInstance().
* Checked some todos in the README.MD file.
* Fixed some bugs and improvements.

!== KEEP IN MIND / WARNING !== 
This version is currently untested, so expect things to be broken. As this is an ALPHA version.
If you have any issues to report, then please open a new issue.
Thank you!