!== /customquery help, more todos !==
Added /customquery help - Shows the command usages. Not all of it is on there, but the main commands are there. Simply just copy the config file options / names into /customquery <setting> <value> So for example, if I wanted fake-slots, simply copy and paste that option to /customquery fake-slots 10, and that'll add the minimum of fake slots. This will update almost instantly without the need of a config reload or server restart.
Oops, I didn't mean v1.0.3 ALPHA. This is v1.0.2 ALPHA - Why am I getting a-head of myself? ;P
Add more to-dos to README.MD
Add some to-dos to the main class file.