Online support (Untested)
This update provides support for Online support. If a player is online, the server will schedule the repeating task for ServerName, not just for offline players support. This also supports online players support as well. (Untested)
There could be bugs with this build. So be cautions before using this build. Thank you!